+++ src/executor_v2/explain.rs
@@ -22,6 +22,7 @@ impl ExplainExecutor {
         let explainer = binding.pretty();
         let mut explain = String::with_capacity(1000);
         let mut config = PrettyConfig::default();
+        config.need_boundaries = false;
         config.unicode(&mut explain, &explainer);
         let chunk =
             DataChunk::from_iter([ArrayImpl::new_utf8(Utf8Array::from_iter([Some(explain)]))]);
